COMP: Fix compilation warnings and typo
* unused variable
* signed vs unsigned comparison
* old copy paste settings label

Closes #1970

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=19985
git-svn-id: http://svn.slicer.org/Slicer4/trunk@19985 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLDisplayableNode.cxx
@@ -257,7 +257,7 @@ void vtkMRMLDisplayableNode::UpdateDisplayNodes()
 //----------------------------------------------------------------------------
 void vtkMRMLDisplayableNode::UpdateNthDisplayNode(int n)
 {
-  assert(n >= 0 && n < this->DisplayNodeIDs.size());
+  assert(n >= 0 && n < static_cast<int>(this->DisplayNodeIDs.size()));
   vtkMRMLNode* node = this->GetScene() ?
     this->GetScene()->GetNodeByID(this->DisplayNodeIDs[n]) : 0;
   this->SetAndObserveNthDisplayNode(n, vtkMRMLDisplayNode::SafeDownCast(node));
